Handle re-ingests of deleted packages correctly
bpa-ingest creates consistent package identifiers during the ingest
process.

If the data is removed from the mirror and bpa-ingest is run with the
delete flag, bpa-ingest uses CKAN API package_delete call to make the
package no longer visible on the portal and mark it as deleted.
(effectively "state": "deleted")

We are not able to recreate this package (with our consistent
identifier) until it is fully purged (either via the admin interface, or
via the dataset_purge CKAN API call)

This commit fixes the behaviour of the bpa-ingest to purge any deleted
packages first to enable them to be recreated (under the assumption that
since we are reingesting the data this is safe to do)