Add resource_path to resource metadata to complement base_url
Each package has a base_url attached to it.   This reflects the location
of all the resources and metadata on the BPA mirror server.

Some datatypes have the individual resource files in directories
below that base_url.

We need a "resource_path" to be able to reassemble the full URL to
the resource on the BPA mirror with the metadata associated with
the package/resource on the BPA Data Portal.

See: https://trello.com/c/NjWO3SPT/1943-add-additional-metadata-to-resources-to-retrieve-the-mirror-url